backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
pypdf
sentence-transformers
pinecone